bloodsmearimageanalysisproject/project/backend/requirements.txt
torch
torchvision
transformers
flask
flask-cors
pymongo
pillow
numpy
gunicorn